refactor: separate sprite comp
